Update Gravatar.php
Sometimes I code without an internet connection.
When this happens, the package returns an error due to the get_headers() PHP function, inside exists() function.

Adding a @ before function, exists() will return FALSE whenever the user is not connected to the internet. So solving the problem.

What do you think?